add folders to Xcode project, move conversationCellModel to separate file, removed old main.storyboard, ViewController
