+++ T-Chat/ConversationCell.swift
@@ -24,14 +24,6 @@ class ConversationCell: UITableViewCell {
     
 }
 
-struct ConversationCellModel {
-    let name: String
-    let message: String
-    let date: Date
-    let isOnline: Bool
-    let hasUnreadedMessages: Bool
-}
-
 extension ConversationCell: ConfigurableView {
     
     typealias ConfigurationModel = ConversationCellModel
