+++ T-Chat/ConversationCellModel.swift
@@ -1,0 +1,17 @@
+//
+//  ConversationCellModel.swift
+//  T-Chat
+//
+//  Created by Дмитрий on 29.09.2020.
+//  Copyright © 2020 DP. All rights reserved.
+//
+
+import Foundation
+
+struct ConversationCellModel {
+    let name: String
+    let message: String
+    let date: Date
+    let isOnline: Bool
+    let hasUnreadedMessages: Bool
+}
